feat: expose "unwrap" via "using for ... global"
feat: add "intoInt256" in "SD59x18"
feat: add "wrap" and "unwrap" in "SD1x18" and "UD2x18"
fix: check for overflow in "SD59x18_IntoUD2x18"
fix: check for underflow in "SD59x18_IntoSD1x18"
docs: improve wording in NatSpec code comments
refactor: change type of "MAX_UINT128" to "uint128"
refactor: change type of "MAX_UINT40" to "uint40"
refactor: revert with specific custom error in "SD59x18_IntoUD60x18"
refactor: simplify casting to "SD1x18" and "UD2x18"
test: add "MAX_UINT128" and "MAX_UINT40" constants in "BaseTest"
test: inherit from "PRBMathUtils" in "BaseTest"
test: refactor "test_RevertWhen" to "testFuzz_RevertWhen"
test: remove value type prefix from helpers and casting file names
test: rename "mathematical" directory to just "math"
test: test all casting functions available for "SD59x18"
test: test all casting functions available for "UD60x18"
test: test "unwrap" and "unwrap" for "SD1x18"
test: test "unwrap" and "unwrap" for "UD2x18"
test: use "x.unwrap" instead of "unwrap(x)"

+++ src/Common.sol
@@ -25,10 +25,10 @@ error PRBMath_MulDivSigned_Overflow(int256 x, int256 y);
 //////////////////////////////////////////////////////////////////////////*/
 
 /// @dev The maximum value an uint128 number can have.
-uint256 constant MAX_UINT128 = type(uint128).max;
+uint128 constant MAX_UINT128 = type(uint128).max;
 
 /// @dev The maximum value an uint40 number can have.
-uint256 constant MAX_UINT40 = type(uint40).max;
+uint40 constant MAX_UINT40 = type(uint40).max;
 
 /// @dev How many trailing decimals can be represented.
 uint256 constant UNIT = 1e18;
